2.15.42

+++ binancedelivery-api/src/main/java/io/contek/invoker/binancedelivery/api/websocket/user/AccountUpdateChannel.java
@@ -5,7 +5,6 @@ import io.contek.invoker.binancedelivery.api.websocket.common.WebSocketEventMess
 import javax.annotation.concurrent.Immutable;
 import javax.annotation.concurrent.NotThreadSafe;
 import javax.annotation.concurrent.ThreadSafe;
-import java.math.BigDecimal;
 import java.util.List;
 
 import static io.contek.invoker.binancedelivery.api.websocket.user.constants.UserEventTypeKeys._ACCOUNT_UPDATE;
@@ -52,20 +51,20 @@ public final class AccountUpdateChannel
   public static final class BalanceUpdate {
 
     public String a; // asset
-    public BigDecimal wb; // wallet balance
-    public BigDecimal cw; // cross wallet balance
+    public Double wb; // wallet balance
+    public Double cw; // cross wallet balance
   }
 
   @NotThreadSafe
   public static final class PositionUpdate {
 
     public String s; // symbol
-    public BigDecimal pa; // position amount
-    public BigDecimal ep; // entry price
-    public BigDecimal cr; // pre-fee accumulated realized
-    public BigDecimal up; // unrealized PnL
+    public Double pa; // position amount
+    public Double ep; // entry price
+    public Double cr; // pre-fee accumulated realized
+    public Double up; // unrealized PnL
     public String mt; // margin type
-    public BigDecimal iw; // isolated wallet
+    public Double iw; // isolated wallet
     public String ps; // position side
   }
 }
